King messaging clarity (#8426)
# About the pull request

Gives hints to newer players about how to destroy or protect a growing
xenomorph king.

# Explain why it's good for the game

Helps newbies a bit, and parity with
https://github.com/cmss13-devs/cmss13/pull/7592


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Gives hints to newer players about how to destroy or protect a
growing xenomorph king.
/:cl: