Adds a shotgun choke for the M37A2 (#8277)
# About the pull request

adds a new barrel attachment for the m37a2 shotgun. it's available in
squad prep and requisitions in small quantities.
it changes the shotgun from a secondary utility weapon to more of a CQC
primary, similar-ish to the m39. sprites were done by thwomper.


![image](https://github.com/user-attachments/assets/d9f71fdc-2fd7-4e4a-8818-cc1b0582f4be)


the stat changes are as follows,
negatives are:
1. reduces the damage output of the shotgun by 20%
2. remove any stunning or slowing ability from slugs/buckshot
3. increases the recoil by 2 steps

positives are:
1. increases the max range of shells by 1 tile
2. doubles the fire rate from 0.5 bullets per second to 1 bullet per
second
3. decreases falloff by 30% 
4. decreases bonus projectile scatter by 5 (buckshot is now a much
tighter formation)
5. increases accuracy by 25%
6. increases bullet velocity by 1 step

it also changes the firing sound to a different one, so enemies can
quickly differentiate it from a standard m37.

VIDEO IS OUTDATED BUT ROUGHLY SHOWS THE WEAPON'S FUNCTIONALITY


https://github.com/user-attachments/assets/a216e9d8-06ea-4648-a042-1710faf3c749

DPS TABLE:
| **AMMO** | **MK2** | **M39** | M37 **CHOKE** |
|-|-|-|-|
|**STANDARD**| 176| 226 | 208 |
|**AP**| 132| 173| 96|

it's fairly solid at clearing out soft targets, but flechette rounds lag
behind regular AP ammo. i think the risk of being forced to get close up
is worth the DPS being comparable to the M39. the range on buckshot is
worse than the M39's, but if you can land a solid click the payoff is
much greater. note, i am unsure of touching flechette damage values
currently due to #8048 being in the air for a testmerge, and i believe
it's fine as is when it comes to the choke right now.

some guncode lines were changed to get damage multipliers working for
bonus projectiles, along with a new variable that increases the max
range of projectiles. PBs now inherit damage multipliers too

# Explain why it's good for the game

it's a new, fun way to utilize a primary weapon. the attachments marines
currently have are pretty lackluster, with most of them being bland stat
increases and some of them just straight up being absolute garbage.
hopefully by adding more unique, mechanic-based attachments and tweaking
some of the others we can make the system a lot better and fun. i think
this attachment is a good start

while the numbers might not be all there yet, they can always be
tweaked, and with the lack of stuns/slow i doubt this weapon will be
that oppressive against xenos.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Vile Beggar, thwomper
add: Adds a new shotgun choke attachment to squad prep and requisitions
vendors for the M37A2 shotgun.
/:cl: