Shivas Panic Room Insert Changes (#7922)
# About the pull request

Changes the presets of the survivors that spawn in the panic room insert
on Shivas. Inspired by [this scene from Aliens: Dark
Descent](https://youtu.be/QMCG9A80L50&t=2074)


# Explain why it's good for the game

When I asked people what their least favorite nightmare insert was, this
one was the most common answer. The most common complaints were that it
was boring and could never spawn engineers. This is an attempt to give
it more sovl and adding a chance for an engineer, with a guarantee of an
engineer if you have at least 5 survivors.


# Testing Photographs and Procedure
<details>
<summary>Screenshot</summary>


![image](https://github.com/user-attachments/assets/62ec0177-e1b9-4c15-8dc2-a82c712a6df4)


</details>


# Changelog
:cl:
maptweak: changed the presets on the shiva's panic room insert to have
an engineer and 1 pmc
/:cl: