Adds French (#7989)
# About the pull request

Adds French as a language selectable in language prefs

also changes monkey language key to _ because there are only so many
numbers.

regarding the language blurb, according to canon french is independent
in the ICSC so its included in the desc as a cool bit of world-building
ig
# Explain why it's good for the game

Its not a niche language and a lot of people's chars speak it, but its
not in the game. Plus more languages is always good.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Adds French as a speakable language.
/:cl: