Allows ghosts to read faxes. (#7938)
# About the pull request

As title, this allows ghosts to view faxes sent throughout the round.

# Explain why it's good for the game

Sometimes it gets boring as an observer, and considering that ghosts can
see faxes briefly when they're sent in the first place, I see no reason
not to allow that to persist.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Ghosts can now view faxes sent through the round.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>