Adds M4A3 AP mags to squad vendor (#7858)
# About the pull request

Adds HP and AP mags for M4A3 to squad vendors.
Vendor pop scaling is the same as speedloaders.

# Explain why it's good for the game

- Special mags for M4A3 are very inaccessible unless you are a SO or IO,
even then you have a very limited amount of mags for it.
- Other guns dont have a lack of special mags in the squad vendors and
because of this, M4A3 is seen as a weak pistol and is mostly not used.

# Testing Photographs and Procedure



# Changelog
:cl: Ansekishoku
balance: adds AP mags for M4A3 pistol to the squad vendors.
/:cl: