hybrisa light stumps (#8445)
# About the pull request

This PR is a follow up to #6248 making it so Hybrisa street lamps and
traffic lights when hit by a > EXPLOSION_THRESHOLD_LOW `ex_act` will
deconstruct into a stump. The stump has a 10% chance to knock a crossing
human over when passing but does not have the shimmy behavior.

The var removals are just removal of overrides that didn't change
anything.

# Explain why it's good for the game

Normally I don't want colony lights destructible, but this was
frequently requested by marine players, so if they think the loss of
light is worth avoiding an obstacle, they can have it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![stump](https://github.com/user-attachments/assets/2d2f6374-351b-4037-874e-6658e9e27d75)

![image](https://github.com/user-attachments/assets/8963e5d1-7d51-4d39-9dc5-c254e26d7bed)


</details>


# Changelog
:cl: Drathek Zenith
add: Colony lights and traffic lights can be exploded into a stump now
imageadd: Lightpost stumps by Zenith
/:cl: