Field Doctor (#7857)
# About the pull request


Just like the PR that split Pilot Officers, this PR aims to make a Field
Doctor occupation, so whoever rolls it gets to deploy instead of arguing
between who wants to deploy.

Doctor and Field Doctor are identical, one has no advantage over the
other and they share the same skills.

There is a 5 hour requirement on medical roles to be a Field Doctor
right now, rather than a jokingly low 1 hour on normal Doctor role
because it needs more skill and knowledge. I can lower this if this is a
problem, though i think it is within reason.
I also might need some suggestions on the Field Doctor's shortened job
name when they talk over comms, if the current one is a problem that is.


# Explain why it's good for the game

Let’s say 4 people roll Doctor. Three of them want to deploy, but only
one can be deployed—there’s no other option.
The others who lose rock paper scissors or whatever will go cryo and
leave.

This occupies the Doctor job slots until the time it's decided who will
be deployed, but by that time, most people decided what they want to
play and joined the game. This restricts people who actually want to
play as doctors from playing the job.

The purpose of this PR is to ensure that those who genuinely want to
play as doctors get to play as doctors.

People who want to deploy roll Field Doctor, if they get the roll, they
deploy; if they don’t, they won’t.
This way, the doctors that normally would cryo after learning that
someone else is deploying will have the chance to play other roles
instead, because they didnt join the round yet and can choose to play
something else.



-Being a FOB doctor requires more in game knowledge and skill.
-How it usually goes is that either the FOB doctor has good knowledge on
what they are doing OR knows nothing and ruins the round of the person
who just came to get some fracs fixed and come out with more fracs, an
OD, and organ damage.

-It makes no sense that whoever gets to ask for permission to the CMO
first gets to deploy, it's just a race not determined by a fair job
roll.
-Or even worse, they argue over the deployment and it usually ends in a
grudge between two people, and other doctors that are not interested in
being a shipside doctor goes cryoing, leaving the almayer with no
doctor.


# Testing Photographs and Procedure

Job and Minimap icons



![image](https://github.com/user-attachments/assets/e8ecfbc1-9f84-418c-81a1-16a46515332d)

![image](https://github.com/user-attachments/assets/0b97ec98-a4f2-44e0-aece-ac3b49a8d15f)



# Changelog
:cl: Ansekishoku
add: A new Field Doctor job to end the arguments between doctors about
who wants to deploy.
/:cl:

---------

Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>