Prevents C4 being placed on explosion proof atoms. (#8510)
# About the pull request

As title

# Explain why it's good for the game

Makes no sense to be able to put explosives onto an object that cannot
be harmed by explosives.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: You can no longer place C4 or Breaching Charges onto objects that
are explosion proof.
/:cl:

---------

Co-authored-by: forest2001 <forest2001@cm-ss13.com>