Acid Runner: Acid gen cap (#7850)
# About the pull request

* Adds a 400 acid cap on acid passive generation. This cap stops passive
acid generation once it is met.
* Adds additional acid generation beyond the cap during active combat.
This is done by turning on additional passive acid generation for a few
seconds every time the player slashes or tail stabs an enemy. If the
player is under the cap, this stacks over a normal passive generation.
The acid cap itself does not influence if this can be triggered or not,
so it can be triggered ABOVE the cap.
* Adds hud element to show acid generation cap being reached. This is
the blue dot.
* Adds status panel element to show current max acid gen cap
* Adds status panel element to show if combat acid gen is active or not
* Changes status panel to show the maximum acid next to the current acid
* Adds UI element to show if combat acid is active or not.

# Explain why it's good for the game

Acid Runner has been controversial for a while now.
The ability " For the hive" is mainly the problem. It surrounds many of
the same problems OT has, such as the fact that you can gain something
immensely powerful with little to no effort. But outright removing it
isn't the best because "For the Hive" is very fun for the player using
it.
So let's make it just fairer to get, instead of nerfing the ability
itself.

Currently, it's possible to get the highest possible explosion by doing
nothing.
The cap on passive generation changes that.
It stops once the threshold is met, and starts again once acid gets
lower.

Furthermore, there is now COMBAT ACID GENERATION!!!!
This means you gain 1 acid every tick while you fighting.
This means players have 1:1 the same acid gain while they are in combat,
pre-cap.
Furthermore, it stacks ONTOP of the passive acid generation. This means
players can reach the cap much faster if they engage in combat. This
should heavily reward acid runners to be more active now while keeping
the overall acid gain for active players pretty much unchanged.

The cap should be high enough though for more inexperienced players to
still get enough acid. So even players that aren't that good, still have
the opportunity to use Acid Runners full kit.
If 400 turns out to low, it can easily be changed. Only one value needs
to be changed, and the visual indicators for it follow suit. No problems
there.

(don't forget to ping me on Discord if you wanna talk balance, NHC)

Thanks to Clairion for giving me the idea of combat acid.
And BIG thanks to Awen who was so kind to guide me through coding
problems for hours, and taught me many things during it.

# Testing Photographs and Procedure
<details>


![grafik](https://github.com/user-attachments/assets/a17b07e2-d6ec-4944-84f8-38d73360ee0f)


![grafik](https://github.com/user-attachments/assets/28ef7b99-f242-468e-9c7e-7ad7daf8d11e)


![grafik](https://github.com/user-attachments/assets/ae9578a2-f286-41e8-bee5-ed854a89f484)

https://gyazo.com/87ec5a1172540be9f1dbdb132ef16d0c

<summary>Screenshots & Videos</summary>

</details>


# Changelog
:cl: NHC
add: Added a passive acid generation cap of 400 to acider runner
add: Added an additional passive acid generation that activates when in
combat for acider
ui: Added status panel info to show if combat acid generation is active
or not for acider
ui: Updated status panel info to show acid gen cap and max acid for
acider
ui: The tgui_alert modal now calculates height differently to better
handle large messages
ui: Purchase Strain buttons are now "Mutate" and "Cancel" so they don't
stack in the modal
imageadd: Added overlays for caps for 10, 20, 30, 40, 50, 60, 70, 80 and
90%
imageadd: Added overlays showing acider drooling when combat acid
generation is active
spellcheck: Revised acider purchase strain information
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>
Co-authored-by: Drulikar <drathek@gmail.com>