Nightmare Insert: Hybrisa - IASF Outpost Souter (#7954)
# About the pull request

Adds the Imperial Army Space Force Outpost Souter to Hybrisa Prospera.
In-lore, this is intended to represent the period of time prior to the
withdrawal of the IASF from the colony due to political pressure from
the Company.

Outpost Souter is a smaller, ancillary, basing for a single UD4-UK
Dropship and a small detachment of ISAF personnel. The primary, and
larger, base would be further west of the colony, as represented by the
(blocked off) escape tunnel network under the base.

This insert will over right the Company Offices, located outside of the
space port. The concept being that the Company knocked down the outpost
and replaced it with an office building.

The insert is currently set to occur at a 25% chance, this can be
changed if desired.

Furthermore, this insert does not add new IASF survivors or corpse
spawners. This would require the work of a spriter, and would be a
future project if anyone wants to pursue it.

# Explain why it's good for the game

More variation on a map, gives more backstory hints as well. 


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![twe](https://github.com/user-attachments/assets/ee18319d-6aeb-4e96-8564-d3017a171fdd)


</details>


# Changelog
:cl:
mapadd: IASF Outpost Souter has been added as a nightmare insert for
Hybrisa Prospera. Part of the larger detachment of Imperial Armed Space
Forces that defended the colony. The IASF had been forced to withdraw
prior to the events of the xenomorph outbreak, however this insert
represents a timeline where the IASF had not fully completed their
withdrawal.
code: Added a set of transparent versions of the TWE cosmetic Dropship
for mappers to use, instead of the older version that leaves a void
overlay.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>