Added missing M42A Urban backsprite icons (#8560)
# About the pull request

Added missing back sprites for the Urban variant of the M42A. For some
reason, it was left blank in guns_by_map/urban, which led to it being
invisible when placed on the backpack slot on Hybrisa.

# Explain why it's good for the game

No invisible guns + drippy urban sniper now appears on your back
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/e68af389-c7d0-432c-825f-6a46719c8cac)

![image](https://github.com/user-attachments/assets/c43e5f28-9796-41b0-895a-40ff542015dd)

![image](https://github.com/user-attachments/assets/aec6177f-67b8-4206-8f09-e7b43e3f8859)

![image](https://github.com/user-attachments/assets/90afb603-84d5-4610-9a19-5ffaccb3faa0)


</details>


# Changelog
:cl:
imageadd: Added missing back sprites for the M42A's Urban variant.
/:cl: