Xeno overwatch QOL (#8562)
# About the pull request

changes the following things:

1. trying to overwatch a xeno you are already watching will make you
stop watching themo
2. the hive tracker will now overwatch whatever leader you're currently
tracking. trying to overwatch a tunnel will make you overwatch the queen
instead.

# Explain why it's good for the game

it's just a little bit of QOL. the watch button making you exit is
something i'd expect it to already do, but it doesn't and it forces you
to either click the ability button to exit or move around which is kind
of annoying.
the hive tracker change makes personally makes sense to me, because if a
player has it set to a leader they're going to be curious about where
and what they're doing, more so than what the queen is doing (she's
probably on ovi if you're choosing to track a leader)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>




https://github.com/user-attachments/assets/73e8358b-8557-4e71-a4b1-7deb6f7e1449







</details>


# Changelog
:cl:
qol: Clicking 'Watch' while already overwatching a xeno will now make
you exit overwatch. You can now overwatch leader xenos with the hive
tracker, instead of only being able to see the Queen.
/:cl: