Lets COs (or XOs if they're brave) fire their protag sidearms into the air (#8558)
# About the pull request
Adds a new unique ability to the Mateba and deagle that lets it be fired
into the air like the flare gun. Does NOT replace spinning the barrel,
and can only be used on grab or harm intent.

# Explain why it's good for the game

Effective method of crowd control


# Testing Photographs and Procedure

https://github.com/user-attachments/assets/f064e102-5ebc-45eb-bbce-06bed8e7c3a0
# Changelog
:cl:
add: Mateba and CO Deagle unique ability 'fire into the air'
/:cl: