Kutjevo Synth Survivors (#8557)
# About the pull request

This adds unique synthetic survivors to Kutjevo Refinery, as well as a
subtype of PMC synth armor that uses the sprite of the goon vest and has
one less storage slot, making it a direct downgrade from its higher
quality counterpart. This subtype is only used for the corporate
security synth survivor in this PR and nowhere else. Haven't done this
in a while and I wouldn't be surprised if I made a few errors while
writing this all, so I apologize in advance.

# Explain why it's good for the game

Having consistency with the survivors on a map is good, Kutjevo's
survivors all have the map-specific standardized jumpsuits and
helmet/goggle/mask combos, so the colony synthetics should also have
these standardized outfits. For the most part I've tried to make them
standard Kutjevo jumpsuits with a few exceptions where I thought the
sprites fit well and added a bit of variety to the loadouts. Some of
them may look same-y, but that's kind of the point. All of the regular
human survivors also mostly look the same on account of the standard
uniform.

I think the Corporate Security synth having the unique vest isn't an
earth-shattering balance change, it's essentially a straight downgrade
from just having a hazard vest or external webbing except for the fact
it has a mounted light and can hold a general utility pouch. It's also
not on par with its higher grade counterpart since it has one less
storage slot. I intend for this to be used here and not for any ERTs or
anything. Fits with the map's scattered goon corpses and the goon
survivors that can also spawn.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![synths](https://github.com/user-attachments/assets/e16bbdb8-acbd-462c-a7a3-eee43cfaec6c)

Standard civilian, 2 engineer variants, medical, scientist, corporate
security, liaison's assistant, CMB synth.

</details>


# Changelog

:cl: IowaPotatoFarmer
add: Added unique map-specific synthetic survivors to Kutjevo Refinery.
add: Added a lower quality goon squad synthetic vest, only used by the
Kutjevo Refinery corporate security synth for now.
/:cl: