Synth key can now be used with engineering pamphlet training, as intended. (#8520)
# About the pull request

Reduces skill to use synth keys to engineering novice, which is the
level of engineering granted by pamphlets. This seems to have been one
of the intents behind @Blundir 's PR #8127 "balance: change synth key
skill level from engi to trained (you can use key after reading engi
pamphlet)" but the level set was wrong.

# Explain why it's good for the game

As is, there is no way for a medic to ever be able to revive synthetics.
This, I believe, was intended(?) to be changed by the original PR.

Instead of reducing the price or giving medics both the ability to use
and the means to acquire reset keys, this is an intermediate. Medics
(and PFCs, I suppose) who have read the engineering pamphlet can now
revive the synth - PROVIDED they can source a key. I don't foresee this
being too much of an issue because you need:

1: Someone who has read a pamphlet
2: Someone who actually cares enough to use the pamphlet skill to revive
the synth (a lot of PFCs probably wont, let's be honest)
3: The synth still needs to be repaired (Given the use case for engi
pamphlets I usually see is just to put cades up, most people w/ them
lack tools + welding goggles so enjoy being turbo-blind)
4: The key itself (most synths do carry keys somewhere on them, so this
is the softest of the reasons. Marines still need to root through their
shit to find it and then have to hope it's not been used)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
balance: the skill required to use a synth reset key has been reduced to
novice (you can use it w/ a pamphlet)
/:cl: