Update to Lower Medbay - Complete redesign (#8292)
# About the pull request

This PR focuses on updating the lower medbay in a way that the
"long-hallway" concept becomes a thing of the past, with lots of new
features and quality-of-life improvements included, while basically
maintaining the exact same size as the original lower med.
# Explain why it's good for the game

This re-design have been going through my personal development have more
than a year now, and given my attempts on all kinds of design drawings,
I choose this to be the most efficient, quality-of-life improving, and
in overall better than the original counterpart. Everything passed
through detailed thought, and this is the result of the refinement.


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/0f01c3db-dd39-48d1-9154-03df55e0ea47)

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Updates Lower Medbay to uh... *different* standards. Updates
morgue to allow for more symmetry. Changes the CMO's bunks a little to
accomodate the new set of ladders. Updates research to have more
internal space in the containment area, including other addons.
add: Adds an access define for Field Medical supply room.
/:cl:

---------

Co-authored-by: Nanu308 <59782240+Nanu308@users.noreply.github.com>
Co-authored-by: forest2001 <forest2001@cm-ss13.com>