Flashing Lights Pref Toggle (#8575)
# About the pull request

This PR adds a preference to opt in (as in flashing effects are disabled
by default) to flashing effects like the AR goggles and the happy pills.
You can set this both in character preferences and the toggle
preferences menu.

I also autodoc'd most of the defines for client prefs.

# Explain why it's good for the game

More preferences for something that suddenly is contentious.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/f2526773-243f-4ec0-8b3c-e987f9053a3b)

![disco](https://github.com/user-attachments/assets/b3951017-96df-49b5-8378-97dd57187ac5)

</details>


# Changelog
:cl: Drathek
add: Added a preference to opt in for flashing lights (e.g. AR goggles
and happy pills)
code: Cleanup client_prefs
/:cl: