Fixes APS (again) (#8597)
# About the pull request

Anti para has been broken, whenever it regressed the stage of an embryo
it would set the progress counter specificaly back to the progress limit
for a stage, which is a problem because the next time the embryo
progress ticked it would see that the progress limit has been reached,
thus increasing the stage, then anti para would come around again and
decrease the stage.
Thus creating a back and forth of stepping between the two stages untill
the APS ran out.

We fix this by essentially taking an extra tick's worth of progress off
the new stage.

# Explain why it's good for the game

Xeno benefitting bugs bad.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Antiparasitic chems can cure stage 2+ infected again.
/:cl: