Removes duplicate MP belt line in SecTech (#8586)
# About the pull request
closes #6689
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Duplicates bad.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

![image](https://github.com/user-attachments/assets/e15b2bf3-c49b-4d34-a883-0e37c14344f6)

</details>


# Changelog
:cl:
fix: Fixes setting the amount of belts in the SecTech twice.
/:cl: