Create-Xenos range defaults to 0, instead of 1 (#8582)
# About the pull request
when Geeves made the create-xenos menu, it was set to 1, instead of
being 0, like the create-humans menu. This has caused extreme emotional
pain whenever I try to spawn xenos in a line and one spawns diagonally
and I hate it I hate it I hate it

So now the default is 0 range, just like create-humans.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Causes less pain to anyone trying to spawn xenos using the menu
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
admin: Create-Xenos range defaults to 0, instead of 1, hooray!
/:cl: