Light Replacer Improvements (#8509)
# About the pull request

Revival of #7381. Adds Light Replacer to ComTech and normal tool storage
vendors, adds recipe to Autolathe. Recycles broken lightbulbs, both when
replacing them in holders or inserting them manually, turning 3 broken
bulbs into 1 new light.

# Explain why it's good for the game

Light Replacers are handy but understated tools for quickly getting
light back to places with active APCs but is barely avaliable outside of
ones placed on maps and isn't as smooth to use as it could be.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added the Light Replacer to tool storage and to the autolathe. The
Light Replacer can now also recycle broken bulbs, either directly when
replacing or when inserted by hand.
/:cl:

---------

Co-authored-by: harry <me@harryob.live>