some Tgui fixes (#8652)
# About the pull request

There's a rare case where the chat might not init, this also adds a
second check if telemetry was received and if not, re-requests it once.

Fixing some tgui issues.

recommending to TM.

Brings the chat closer in sync with
https://github.com/tgstation/tgstation again, some of the other changes
that exist on TG are likely better for you to check on your own,
especially the option to link / unlink font sizes, etc.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: numberInput works properly again with mouse movement
fix: the chat will no longer init before settings are loaded
qol: chat settings can be in / exported
qol: chat tabs can be reordered
/:cl: