Removes devour, replaces it with haul. #2737 Revival (#8442)
# About the pull request

Replaces xeno devour with hauling, while keeping it as mechanically
close as possible to the previous mechanic. Also removes
stomach_contents from code

Also Fixes #8382
Technically Fixes #7966 
Fixes #8418 

# Explain why it's good for the game

I was told devour isn't lore accurate. I don't think xenomorphs ever
actually swallowed hosts whole like pythons in the movies, although I
actually never watched any of them so I couldn't tell you. Also there's
the other obvious reason

~While hauling functions basically the same as devour (host won't take
damage from external sources while hauled, etc), the damage taken by the
xenomorph while a host is trying to break out of the haul might be lower
or higher than with devour, as it used to be a 30% chance of dealing
damage to xenomorph when trying to move inside of them, whereas now you
will just hit the xenomorph with your weapon with a regular melee attack
cooldown. I tried to adjust the damage accordingly, but it may need to
be raised or lowered in testing.~

Breaking out of the haul now works the same and deals the same damage
like previous devour

Credit towards Molto for the original PR this works off on, as well as
Kugamo which I was told by Molto that he helped a lot with the original
PR.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

![image](https://github.com/user-attachments/assets/d9529632-0f4d-4df6-8f73-3250d4bbba3a)


https://github.com/user-attachments/assets/1899fdcd-d004-45c0-a220-519adfb54a35



</details>


# Changelog
:cl: Mikrel0712, Molto, Kugamo
add: Replaced xenomorph devour with hauling.
/:cl:

---------

Co-authored-by: casperr04 <kacperjaroslawurbanczyk@gmail.com>
Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>
Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>