Remaps and Adds a Fax Machine to Whiskey Outpost CIC. (#8476)
# About the pull request

Remaps WO CIC to be a bit wider, remaps Synth Prep to not be weirdly
oblong with jagged wall edges.
Also adds a Fax Machine to CIC.

# Explain why it's good for the game

Fax Responder Content and CIC RP potential. Prior to this, going Fax
Responder was basically useless unless you had a Combat Correspondent
and even that can be a slog. So I made room in CIC for a USCM Fax
Machine that will allow the XO and CO to fax High Command if need be.
(And whatever other factions.)


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
mapadd: Whiskey Outpost CIC has been widened, and a USCM Fax Machine
added.
mapadd: Whiskey Outpost Synth Prep has been remapped to be more
streamlined.
/:cl: