Fixes the non-existence of the fax responder listening bugs again (#8515)
# About the pull request

I have fixed this a few times but I'm too much of an idiot to keep it
fixed.

# Explain why it's good for the game

Makes the listening devices I added actually used - yet again.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Re-re-re-re-added the fax responder listening bugs to the USS
Almayer.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>