Refactor urange (#8677)
# About the pull request

This PR refactor urange, splitting it into two procs: long_range and
long_orange. It removes the behavior of adding an area to the list
(nothing used this aspect), and fixes urange with dist 0 not including
center contents if it was a turf.

# Explain why it's good for the game

Ultimately this was done because of confusion about what the proc does
and why. That should now be clarified, and it should be the tiniest bit
more performant.

# Testing Photographs and Procedure
See comment below.

# Changelog
:cl: Drathek
refactor: Refactored urange into long_range and long_orange
/:cl: