K9 Tracker Bug Fix (#8690)
# About the pull request

The active state of the K9 tracker was removed during a reorg of the
sprite files. Placing it back to fix the invisible tracker bug.

# Explain why it's good for the game

Bug Fix Good


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: fixed missing sprite for active K9 tracker
/:cl: