Fixed prop gun wield message showing incorrect name (#8743)
# About the pull request

Fixes Issue #8010  

The incorrect chat message was because place_offhand(user,
initial(name)) was returning the original name of the prop gun as
defined in its prototype ("prop gun") instead of the updated name. This
caused the wielding message to print "You grab prop gun with both
hands.". By replacing initial(name) with name, the updated name (in this
case "Broken M41A pulse rifle") is passed to the place_offhand()
function to be printed. I moved the addition of " (Wielded)" to the name
variable after the place_offhand() call to stop the extra text from
being included in the chat message, which seems to be the original
intention of using initial(name). I was unable to test the changes in a
live environment due to the factors discussed in the testing section of
this PR, but the game compiles and runs.

Relevant files:
code\game\objects\effects\spawners\prop_gun_spawner.dm
code\game\objects\items\weapons\twohanded.dm (modified by this PR)


# Explain why it's good for the game

Addresses an inaccuracy in chat messages when wielding the item. This is
a small bug fix.


# Testing Photographs and Procedure

I confirmed that the code compiles and the game runs with the Dream
Daemon, but was unable to find the hanger referenced in the issue in
order to test the changes. I tried searching the wiki, but couldn't find
any information on the item or hanger. I wrote a test script although I
am unsure if it works properly (I am new to DM and this game), and was
curious if you guys had any advice for me on this.


# Changelog

:cl:
fix: Fixed a bug that caused the Broken M41A to be called Prop Gun in a
chat message.
/:cl:

Co-authored-by: Nicholas Wooldridge <wooldrni@oregonstate.edu>