Hostile Survivor Rework Revival (#7816)
# About the pull request

Revival of #6308. Credit to Steelpoint and Drathek for 99 percent of the
code.

Survivor players can no longer roll for a Hostile Survivor job role.

All maps, with the exception of Chances Claim, have an ~10% chance to
spawn all survivors as CLF operatives onboard a crashed CLF ship. (or
something else)

The LV-624 CLF spawn has had its spawn odds increased.

awesome new van sprite from dimdimich1996

# Explain why it's good for the game

The current iteration of hostile survivor is a role that is detrimental
to the survivor experience. They typically have an extraordinarily low
impact on the round, are often likely to abandon the regular survivors
to try to hide (thus depriving the survivors of vital aid), and most of
the time are killed by the Marines with no real impact.

My proposed change adds a new dimension to the survivor and groundside
gameplay loop. Marines can no longer be 100% certain that the survivors
are all going to be neutral (and xenomorphs who now have a risk to
contend with more organised and armed survivors). This will also boost
survivor cooperation in all scenarios, as regular survivors won't have
to worry about a CLF who runs away, and if the survivors spawn as CLF,
they all now work towards a common goal.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Steelpoint, Drathek, bobama, dimdimich1996
del: Survivors can no longer select the Hostile Survivor job role.
add: All maps, except for Chance's Claim, have an approximately 10%
chance to spawn all survivors as hostile survivors.
add: All players are automatically opted-in to being a hostile survivor
by default.
add: Players can opt-out of being selected for a hostile survivor spawn,
or to only roll for a hostile survivor and not a regular survivor, via
your Personal Preferences settings page.
maptweak: All maps that feature hostile survivors have a spawn area for
the CLF, that spawns alongside the hostile survivors.
balance: The odds of the CLF crashed ship spawn on LV-624 occurring has
been increased slightly per this.
fix: Fixed trolly vehicle hard deletes
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: Steelpoint <6595389+Steelpoint@users.noreply.github.com>
Co-authored-by: Drulikar <drathek@gmail.com>
Co-authored-by: Nanu308 <59782240+Nanu308@users.noreply.github.com>